refectored capabilities into script_parser
